# gzip compression for requests