--- a/Cargo.toml
+++ b/Cargo.toml
@@ -34,6 +34,7 @@ metrics = ["opentelemetry/metrics", "ureq"]
 async-trait = "0.1"
 bytes = "1"
 chrono = "0.4"
+flate2 = "1"
 http = "0.2"
 once_cell = "1"
 opentelemetry = "0.17"