--- a/src/uploader.rs
+++ b/src/uploader.rs
@@ -1,5 +1,7 @@
 use crate::{models::Envelope, Error, HttpClient};
 use bytes::Bytes;
+use flate2::write::GzEncoder;
+use flate2::Compression;
 use http::{Request, Response, Uri};
 use serde::Deserialize;
 #[cfg(feature = "metrics")]
@@ -33,9 +35,10 @@ pub(crate) async fn send(
     endpoint: &Uri,
     items: Vec<Envelope>,
 ) -> Result<(), Error> {
-    let payload = serde_json::to_vec(&items).map_err(Error::UploadSerializeRequest)?;
+    let payload = serialize_request_body(items)?;
     let request = Request::post(endpoint)
         .header(http::header::CONTENT_TYPE, "application/json")
+        .header(http::header::CONTENT_ENCODING, "gzip")
         .body(payload)
         .expect("request should be valid");
 
@@ -50,11 +53,12 @@ pub(crate) async fn send(
 /// Sends a telemetry items to the server.
 #[cfg(feature = "metrics")]
 pub(crate) fn send_sync(endpoint: &Uri, items: Vec<Envelope>) -> Result<(), Error> {
-    let payload = serde_json::to_vec(&items).map_err(Error::UploadSerializeRequest)?;
+    let payload = serialize_request_body(items)?;
 
     // TODO Implement retries
     let response = match ureq::post(&endpoint.to_string())
         .set(http::header::CONTENT_TYPE.as_str(), "application/json")
+        .set(http::header::CONTENT_ENCODING.as_str(), "gzip")
         .send_bytes(&payload)
     {
         Ok(response) => response,
@@ -77,6 +81,12 @@ pub(crate) fn send_sync(endpoint: &Uri, items: Vec<Envelope>) -> Result<(), Erro
             .body(Bytes::from(bytes))
             .map_err(|err| Error::UploadConnection(err.into()))?,
     )
+}
+
+fn serialize_request_body(items: Vec<Envelope>) -> Result<Vec<u8>, Error> {
+    let mut gzip_encoder = GzEncoder::new(Vec::new(), Compression::default());
+    serde_json::to_writer(&mut gzip_encoder, &items).map_err(Error::UploadSerializeRequest)?;
+    gzip_encoder.finish().map_err(Error::UploadCompressRequest)
 }
 
 fn handle_response(response: Response<Bytes>) -> Result<(), Error> {