--- a/src/uploader.rs
+++ b/src/uploader.rs
@@ -1,11 +1,11 @@
 use crate::{models::Envelope, Error, HttpClient};
 use bytes::Bytes;
-use flate2::write::GzEncoder;
-use flate2::Compression;
+use flate2::{write::GzEncoder, Compression};
 use http::{Request, Response, Uri};
 use serde::Deserialize;
 #[cfg(feature = "metrics")]
 use std::io::Read;
+use std::io::Write;
 
 const STATUS_OK: u16 = 200;
 const STATUS_PARTIAL_CONTENT: u16 = 206;
@@ -84,8 +84,16 @@ pub(crate) fn send_sync(endpoint: &Uri, items: Vec<Envelope>) -> Result<(), Erro
 }
 
 fn serialize_request_body(items: Vec<Envelope>) -> Result<Vec<u8>, Error> {
+    // Weirdly gzip_encoder.write_all(serde_json::to_vec()) seems to be faster than
+    // serde_json::to_writer(gzip_encoder). In a local test operating on items that result in
+    // ~13MiB of JSON, this is what I've seen:
+    // gzip_encoder.write_all(serde_json::to_vec()): 159ms
+    // serde_json::to_writer(gzip_encoder):          247ms
+    let serialized = serde_json::to_vec(&items).map_err(Error::UploadSerializeRequest)?;
     let mut gzip_encoder = GzEncoder::new(Vec::new(), Compression::default());
-    serde_json::to_writer(&mut gzip_encoder, &items).map_err(Error::UploadSerializeRequest)?;
+    gzip_encoder
+        .write_all(&serialized)
+        .map_err(Error::UploadCompressRequest)?;
     gzip_encoder.finish().map_err(Error::UploadCompressRequest)
 }