… tests/snapshots/http_requests__traces_batch_async_std.snap
… tests/snapshots/http_requests__traces_batch_tokio.snap
… tests/snapshots/http_requests__traces_simple.snap