See #46 

Seems to work correctly. Tested using:

- `cargo run --example opentelemetry`
- `cargo run --example metrics --features metric`

Todo:

- [x] Do HTTP tests work with this?
- [x] Does this have to be optional (opt-in or opt-out)? We allow custom endpoints? Is it possible they don't support gzip? Is it possible users want to turn off gzip compression to reduce CPU usage?
  - As far as I can tell the .NET SDK turns compression on by default and there is no opt-out. I think fewer configuration options are generally better. Having this always turned on sounds good to me.
- [x] Should we also support gzip responses (`Accept-Encoding: gzip`)? Does Application Insights even support that?
  - I tried calling the App Insights endpoint with `Accept-Encoding: gzip, deflate, br`. It doesn't use it for the response.