@frigus02 2022-03-20
I tried to see if this improves things. Here are some numbers.

First observation: even tiny requests take a few hundred milliseconds. It seems to help a lot to increase the max export batch size. For the tests I used a super large max batch size to really see the effect of the compression:

```
export INSTRUMENTATION_KEY=???
export NUM_ROOT_SPANS=10000
export OTEL_BSP_MAX_QUEUE_SIZE=100000
export OTEL_BSP_MAX_EXPORT_BATCH_SIZE=20000
cargo run --example stress_test
```

| Test | Serialize (JSON + GZIP) | Send | Sent body size | Total runtime |
| --- | --- | --- | --- | --- |
| With compression (level 1) | 2.677s | 6.908s | 1089.28 KiB | 12.000s |
| With compression (level 6) | 3.454s | 7.655s | 784.73 KiB | 12.525s |
| Without compression | 0.604s | 8.130s | 13158.34 KiB | 10.102s |

Compression clearly reduces the body size by a lot (x16). However time spent during compression is bigger than the network time savings for me.

This doesn't seem like a clear win to me. So maybe this has to be opt-in for now.

Side note: There is another possible improvement. Currently serialization & send happen sequentially. It could be helpful to use streams so that the HTTP request can already start while data is being serialized. To make this work, [opentelemetry-http](https://github.com/open-telemetry/opentelemetry-rust/blob/3140cfdb15ae2854586bf6210e4a4d1ddfd11127/opentelemetry-http/src/lib.rs#L55) would have to support that.