@birkmose 2022-03-20
Hmm this is interesting. Using the  [compress_file.rs](https://github.com/rust-lang/flate2-rs/blob/master/examples/compress_file.rs) example in the flate2 repo on my machine it takes 39ms to compress 2MB of JSON data. I did a compression test of 13MB data as well, and that took 238ms. Your example at the lowest compression level appears to take an order of magnitude longer? My tests are run on an AMD 3950X cpu for reference.