@frigus02 2022-03-20
That _is_ interesting. I'm going to test the flate2 example directly later to compare. Maybe I'm doing something stupid here.