@frigus02 2022-03-20
Using compress_file.rs I get:

| Mode | Source len | Target len | Elapsed |
| --- | --- | --- | --- |
| debug | 13474859 | 785183 | 1.270050744s | 
| release | 13474859 | 785183 | 104.022545ms |

I realised I didn't run the tests with `--release` before. I also noticed it's faster to serialize JSON into a `Vec` first before passing it through the gzip encoder. Not sure why that is. Either way, here are new numbers:

| Test | Serialize (JSON + GZIP) | Send | Total runtime |
| --- | --- | --- | --- |
| With compression (level 6) | 153ms | 6.986s | 7.832 |
| Without compression | 41ms | 8.916s | 9.636s |

This looks much more like what I expected.